Fixed wrong dataset called in one part of QA Check.
Changed main script so it now calls the request API script every time.

Commented out the request to input a new legiscan_api_key() since that's something people will want to do once and not repeatedly.

Fixed the viz_partisanship and uncommented it, but offered a sample of a GGPlot that is commented out so that we can change it to whichever party/chamber we want to visualize partisanship in.